fix size_type in sha2 functions

+++ include/boost/hash2/sha2.hpp
@@ -391,7 +391,7 @@ private:
 
 public:
     using result_type = std::array<unsigned char, 28>;
-    using size_type = std::size_t;
+    using size_type = std::uint64_t;
 
     static int const block_size = 64;
 
@@ -468,7 +468,8 @@ private:
 public:
 
     using result_type = std::array<unsigned char, 64>;
-    using size_type = std::size_t;
+    using size_type = std::uint64_t;
+
     using detail::sha2_512_base::update;
 
     static int const block_size = 128;
@@ -545,7 +546,7 @@ private:
 public:
 
     using result_type = std::array<unsigned char, 48>;
-    using size_type = std::size_t;
+    using size_type = std::uint64_t;
 
     static int const block_size = 128;
 
@@ -623,7 +624,7 @@ private:
 public:
 
     using result_type = std::array<unsigned char, 28>;
-    using size_type = std::size_t;
+    using size_type = std::uint64_t;
 
     static int const block_size = 128;
 
@@ -702,7 +703,7 @@ private:
 public:
 
     using result_type = std::array<unsigned char, 32>;
-    using size_type = std::size_t;
+    using size_type = std::uint64_t;
 
     static int const block_size = 128;
